pyproject.toml
[tool.poetry]
name = "hello-world"
version = "0.1.0"
description = ""
authors = ["Dima <pieero47@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
django = "==5.2.9"
celery = ">=5.6.0,<6.0.0"
redis = ">=7.1.0,<8.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
drf-yasg = ">=1.21.11,<2.0.0"
django-redis = ">=6.0.0,<7.0.0"
django-celery-results = ">=2.6.0,<3.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
psycopg2-binary = ">=2.9.11,<3.0.0"
gunicorn = "^23.0.0"
django-cors-headers = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pre-commit = "^4.5.1"
flake8 = "^7.3.0"
black = "^25.12.0"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"